FCT-16897: relocate Header to Pages
